src/requirements.txt
openai>=1.0.0
pytest
python-dotenv
